+++ lib/chart/model/theme/item_theme/bubble/bubble_item_options.dart
@@ -57,7 +57,8 @@ class BubbleItemOptions extends ItemOptions {
         multiValuePadding:
             EdgeInsets.lerp(multiValuePadding, endValue.multiValuePadding, t) ??
                 EdgeInsets.zero,
-        maxBarWidth: lerpDouble(maxBarWidth, endValue.maxBarWidth, t),
+        maxBarWidth:
+            ItemOptions.lerpMaxBarWidth(maxBarWidth, endValue.maxBarWidth, t),
         minBarWidth: lerpDouble(minBarWidth, endValue.minBarWidth, t),
       );
     } else {